Treqsy_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
python-dateutil==2.8.2
email-validator==2.1.0.post1
pydantic==2.5.2
pydantic-settings==2.0.3
alembic==1.12.1
python-slugify==8.0.1
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.1
gunicorn==21.2.0
uvloop==0.17.0
httptools==0.6.0
watchfiles==0.20.0
websockets==12.0
python-magic==0.4.27
boto3==1.28.62
redis==5.0.1
python-memcached==1.59
